yatube_api/requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==2.0.12
cryptography==41.0.7
defusedxml==0.8.0rc2
Django==3.2.16
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.2.0
idna==3.6
iniconfig==2.0.0
install==1.3.5
oauthlib==3.2.2
packaging==23.2
Pillow==9.3.0
pluggy==0.13.1
py==1.11.0
pycparser==2.21
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.26.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
toml==0.10.2
typing_extensions==4.8.0
urllib3==1.26.18
